Support for forcing a specific --extern crate type
As of now, rustc tries to detect what kind of crate it is linking
against based on it's file prefix and extension.
With this patch, rustc offers the possibilty of forcing a specific
crate type in cases the filename is not inline with rustc's
expectations, by --extern kind:name=/path/to/name.ext. `kind` may be
"rlib", "rmeta" or "dylib".